change unsigned int -> uint32_t

+++ Mesh.cpp
@@ -14,7 +14,7 @@ void Mesh3D::add_mesh(const Mesh3D& other)
     this->normals.insert(this->normals.end(), other.normals.begin(), other.normals.end());
     this->st_coordinates.insert(this->st_coordinates.end(), other.st_coordinates.begin(), other.st_coordinates.end());
 
-    for (const size_t& idx : other.indices)
+    for (const uint32_t& idx : other.indices)
         this->indices.push_back(idx + idx_offset);
 }
