+++ Mesh.h
@@ -15,8 +15,8 @@ struct Mesh3D
     void        add_mesh(const Mesh3D& other);
     std::string get_obj() const;
 
-    std::vector<Vec3D>  vertices;
-    std::vector<size_t> indices;
+    std::vector<Vec3D>    vertices;
+    std::vector<uint32_t> indices;
 
     std::vector<Vec3D> normals;
     std::vector<Vec2D> st_coordinates;
